Change abstract `AbstractJob` to interface `Job` to make job implementations more flexible
