+++ service/src/main/java/de/starwit/service/jobs/LineCrossingJob.java
@@ -4,14 +4,24 @@ import java.awt.geom.Line2D;
 import java.awt.geom.Point2D;
 import java.time.Duration;
 import java.time.Instant;
+import java.util.concurrent.Executor;
+import java.util.concurrent.Executors;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import de.starwit.persistence.analytics.entity.Direction;
 import de.starwit.persistence.observatory.entity.ObservationJobEntity;
 import de.starwit.service.sae.SaeDetectionDto;
 
-public class LineCrossingJob extends AbstractJob {
+public class LineCrossingJob implements Job {
+    private final Logger log = LoggerFactory.getLogger(this.getClass());
+
+    private final Executor jobExecutor = Executors.newSingleThreadExecutor();
 
     private static int TARGET_WINDOW_SIZE_SEC = 1;
+
+    private ObservationJobEntity configEntity;
 
     private TrajectoryStore trajectoryStore;
     private Line2D countingLine;
@@ -19,7 +29,7 @@ public class LineCrossingJob extends AbstractJob {
     private LineCrossingObservationListener observationListener;
 
     public LineCrossingJob(ObservationJobEntity configEntity, LineCrossingObservationListener observationListener) {
-        super(configEntity);
+        this.configEntity = configEntity;
         this.observationListener = observationListener;
         this.countingLine = GeometryConverter.lineFrom(this.configEntity);
         this.isGeoReferenced = this.configEntity.getGeoReferenced();
@@ -27,6 +37,15 @@ public class LineCrossingJob extends AbstractJob {
     }
 
     @Override
+    public void pushNewDetection(SaeDetectionDto dto) {
+        jobExecutor.execute(() -> processNewDetection(dto));
+    }
+
+    @Override
+    public ObservationJobEntity getConfigEntity() {
+        return this.configEntity;
+    }
+    
     protected void processNewDetection(SaeDetectionDto dto) {
         log.debug("store size: {}", trajectoryStore.size());
